fix: the files have been moved to the right places
