+++ src/backend/noKeyCloud.Application/Abstractions/Repositories/IFolderRepository.cs
@@ -1,0 +1,11 @@
+﻿using noKeyCloud.Domain.Entities;
+using File = System.IO.File;
+
+namespace noKeyCloud.Application.Abstractions.Repositories;
+
+public interface IFolderRepository
+{
+    
+    Task<Folder?> GetFolderByFolderId(Guid folderId, CancellationToken cancellationToken);
+    
+}
