+++ src/backend/noKeyCloud.Application/Abstractions/Repositories/IUserRepository.cs
@@ -7,4 +7,5 @@ public interface IUserRepository
 {
     Task CreateUser(User user);
     Task<User?> GetUserByUsernameOrEmailAsync(string identifier, CancellationToken cancellationToken);
+    Task<User?> GetUserByUserId(Guid userId, CancellationToken cancellationToken);
 }
