+++ src/backend/noKeyCloud.Application/noKeyCloud.Application.csproj
@@ -15,4 +15,8 @@
       <PackageReference Include="MediatR" Version="14.1.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Features\Files\" />
+    </ItemGroup>
+
 </Project>
