+++ src/shared/noKeyCloud.Contracts/File/CreateFileRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace noKeyCloud.Contracts.Authenticate;
+﻿namespace noKeyCloud.Contracts.File;
 
 public record CreateFileRequest
 (
